requirements.txt
asgiref==3.4.1
boto3==1.20.37
botocore==1.23.37
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
django-phonenumber-field==6.1.0
gunicorn==20.1.0
importlib-metadata==4.10.0
jmespath==0.10.0
Markdown==3.3.6
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
zipp==3.7.0
